Refactor | Calculator 덧뺄셈 메서드 리팩토링

+++ Assets/02. Scripts/TestRunner Practice/Calculator.cs
@@ -2,12 +2,12 @@ using System;
 using UnityEngine;
 public class Calculator
 {
-    public int Add(int a, int b)
+    public long Add(long a, long b)
     {
         return a + b;
     }
 
-    public int Subtract(int a, int b)
+    public long Subtract(long a, long b)
     {
         return a - b;
     }
